pyproject.toml
[project]
name = "timelapse"
version = "0.1.0"
description = "絵画制作過程の写真からInstagram Reels用タイムラプス動画を生成するCLIツール"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "pillow>=10",
    "pillow-heif>=0.16",
    "pydantic>=2",
    "pyyaml>=6",
    "imagehash>=4.3",
]

[project.scripts]
timelapse = "timelapse.cli:app"
timelapse-web = "timelapse_web.cli_entry:main"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-mock>=3",
    "pytest-cov>=5",
    "httpx>=0.27",
]
web = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/timelapse", "src/timelapse_web"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=timelapse --cov-report=term-missing"
markers = [
    "integration: 実際に FFmpeg を起動するエンドツーエンド統合テスト",
]

[tool.coverage.run]
source = ["src/timelapse"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "httpx>=0.27",
]
web = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
]
